refactor: add Finbuckle.MultiTenant.Identity.EntityFrameworkCore.csproj and test project
