+++ src/Finbuckle.MultiTenant.EntityFrameworkCore/Finbuckle.MultiTenant.EntityFrameworkCore.csproj
@@ -12,8 +12,6 @@
 
         <ItemGroup>
             <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="$(TfmMajVer).0.*"/>
-            <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="$(TfmMajVer).0.*"/>
-            <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="$(TfmMajVer).0.*"/>
         </ItemGroup>
     </Target>
