+++ test/Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test/Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net10.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="10.0.0-*"/>
+        <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="10.0.0-*"/>
+        <ProjectReference Include="..\..\src\Finbuckle.MultiTenant.Identity.EntityFrameworkCore\Finbuckle.MultiTenant.Identity.EntityFrameworkCore.csproj"/>
+    </ItemGroup>
+
+</Project>
