+++ test/Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test/TestIdentityDbContext.cs
@@ -2,18 +2,20 @@
 // Refer to the solution LICENSE file for more information.
 
 using Finbuckle.MultiTenant.Abstractions;
+using Finbuckle.MultiTenant.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace Finbuckle.MultiTenant.EntityFrameworkCore.Test.MultiTenantIdentityDbContext;
+namespace Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test;
 
-public class TestIdentityDbContext : EntityFrameworkCore.MultiTenantIdentityDbContext
+public class TestIdentityDbContext : MultiTenantIdentityDbContext
 {
     public TestIdentityDbContext(IMultiTenantContextAccessor multiTenantContextAccessor) : base(
         multiTenantContextAccessor)
     {
     }
 
-    public TestIdentityDbContext(IMultiTenantContextAccessor multiTenantContextAccessor, DbContextOptions options) : base(multiTenantContextAccessor, options)
+    public TestIdentityDbContext(IMultiTenantContextAccessor multiTenantContextAccessor, DbContextOptions options) :
+        base(multiTenantContextAccessor, options)
     {
     }
