+++ test/Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test/TestIdentityDbContextTUser.cs
@@ -1,0 +1,28 @@
+// Copyright Finbuckle LLC, Andrew White, and Contributors.
+// Refer to the solution LICENSE file for more information.
+
+using Finbuckle.MultiTenant.Abstractions;
+using Finbuckle.MultiTenant.Identity.EntityFrameworkCore;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+
+namespace Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test;
+
+public class TestIdentityDbContextTUser : MultiTenantIdentityDbContext<IdentityUser>
+{
+    public TestIdentityDbContextTUser(IMultiTenantContextAccessor multiTenantContextAccessor) : base(
+        multiTenantContextAccessor)
+    {
+    }
+
+    public TestIdentityDbContextTUser(IMultiTenantContextAccessor multiTenantContextAccessor, DbContextOptions options)
+        : base(multiTenantContextAccessor, options)
+    {
+    }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        optionsBuilder.UseSqlite("DataSource=:memory:");
+        base.OnConfiguring(optionsBuilder);
+    }
+}
