+++ test/Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test/TestIdentityDbContextTUserTRole.cs
@@ -1,0 +1,28 @@
+// Copyright Finbuckle LLC, Andrew White, and Contributors.
+// Refer to the solution LICENSE file for more information.
+
+using Finbuckle.MultiTenant.Abstractions;
+using Finbuckle.MultiTenant.Identity.EntityFrameworkCore;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+
+namespace Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test;
+
+public class TestIdentityDbContextTUserTRole : MultiTenantIdentityDbContext<IdentityUser, IdentityRole, string>
+{
+    public TestIdentityDbContextTUserTRole(IMultiTenantContextAccessor multiTenantContextAccessor) : base(
+        multiTenantContextAccessor)
+    {
+    }
+
+    public TestIdentityDbContextTUserTRole(IMultiTenantContextAccessor multiTenantContextAccessor,
+        DbContextOptions options) : base(multiTenantContextAccessor, options)
+    {
+    }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        optionsBuilder.UseSqlite("DataSource=:memory:");
+        base.OnConfiguring(optionsBuilder);
+    }
+}
