--- test/Finbuckle.MultiTenant.EntityFrameworkCore.Test/MultiTenantIdentityDbContext/TestIdentityDbContextTUserTRole.cs
@@ -1,25 +1,0 @@
-// Copyright Finbuckle LLC, Andrew White, and Contributors.
-// Refer to the solution LICENSE file for more information.
-
-using Finbuckle.MultiTenant.Abstractions;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
-
-namespace Finbuckle.MultiTenant.EntityFrameworkCore.Test.MultiTenantIdentityDbContext;
-
-public class TestIdentityDbContextTUserTRole : MultiTenantIdentityDbContext<IdentityUser, IdentityRole, string>
-{
-    public TestIdentityDbContextTUserTRole(IMultiTenantContextAccessor multiTenantContextAccessor) : base(multiTenantContextAccessor)
-    {
-    }
-
-    public TestIdentityDbContextTUserTRole(IMultiTenantContextAccessor multiTenantContextAccessor, DbContextOptions options) : base(multiTenantContextAccessor, options)
-    {
-    }
-
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-    {
-            optionsBuilder.UseSqlite("DataSource=:memory:");
-            base.OnConfiguring(optionsBuilder);
-        }
-}
